feat: ability to set test_run
